refactor: 필요없는 코드 제거 및 클래스 이동
