+++ apps/commerce-api/src/main/java/com/loopers/application/provided/MemberFinder.java
@@ -1,6 +1,6 @@
 package com.loopers.application.provided;
 
-import com.loopers.domain.Member;
+import com.loopers.domain.member.Member;
 
 public interface MemberFinder {
     Member find(Long memberId);
