+++ apps/commerce-api/src/main/java/com/loopers/domain/member/Birthday.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.member;
 
 import java.util.regex.Pattern;
