+++ apps/commerce-api/src/main/java/com/loopers/domain/member/Gender.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.member;
 
 public enum Gender {
     MALE, FEMALE;
