+++ apps/commerce-api/src/main/java/com/loopers/domain/member/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.member;
 
 public class MemberNotFoundException extends RuntimeException {
     public MemberNotFoundException(String message) {}
