+++ apps/commerce-api/src/main/java/com/loopers/domain/member/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.member;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
