+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/MemberRepositoryImpl.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.springframework.stereotype.Repository;
 
 import com.loopers.application.required.MemberRepository;
-import com.loopers.domain.Member;
-import com.loopers.domain.MemberId;
+import com.loopers.domain.member.Member;
+import com.loopers.domain.member.MemberId;
 
 import lombok.RequiredArgsConstructor;
