+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/member/MemberV1ApiController.java
@@ -9,36 +9,35 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.loopers.application.provided.MemberFinder;
 import com.loopers.application.provided.MemberRegister;
-import com.loopers.domain.Member;
-import com.loopers.domain.MemberRegisterRequest;
+import com.loopers.domain.member.Member;
+import com.loopers.domain.member.MemberRegisterRequest;
 import com.loopers.interfaces.api.ApiResponse;
-import com.loopers.interfaces.api.member.dto.MemberInfoResponse;
-import com.loopers.interfaces.api.member.dto.MemberRegisterResponse;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/members")
+@RequestMapping("/api/v1/members")
 public class MemberV1ApiController implements MemberV1ApiSpec {
     private final MemberRegister memberRegister;
     private final MemberFinder memberFinder;
 
     @PostMapping
     @Override
-    public ApiResponse<MemberRegisterResponse> register(@RequestBody @Valid MemberRegisterRequest registerRequest) {
+    public ApiResponse<MemberV1Dto.MemberRegisterResponse> register(@RequestBody @Valid MemberRegisterRequest registerRequest) {
         Member member = memberRegister.register(registerRequest);
 
-        MemberRegisterResponse memberRegisterResponse = MemberRegisterResponse.of(member);
+        MemberV1Dto.MemberRegisterResponse memberRegisterResponse = MemberV1Dto.MemberRegisterResponse.of(member);
         return ApiResponse.success(memberRegisterResponse);
     }
 
     @GetMapping("{memberId}")
     @Override
-    public ApiResponse<MemberInfoResponse> find(@PathVariable Long memberId) {
+    public ApiResponse<MemberV1Dto.MemberInfoResponse> find(@PathVariable Long memberId) {
         Member member = memberFinder.find(memberId);
-        MemberInfoResponse memberInfoResponse = MemberInfoResponse.of(member);
+        MemberV1Dto.MemberInfoResponse memberInfoResponse = MemberV1Dto.MemberInfoResponse.of(member);
         return ApiResponse.success(memberInfoResponse);
     }
 }
