+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/member/MemberV1ApiSpec.java
@@ -1,14 +1,13 @@
 package com.loopers.interfaces.api.member;
 
-import com.loopers.domain.MemberRegisterRequest;
+import com.loopers.domain.member.MemberRegisterRequest;
 import com.loopers.interfaces.api.ApiResponse;
-import com.loopers.interfaces.api.member.dto.MemberInfoResponse;
-import com.loopers.interfaces.api.member.dto.MemberRegisterResponse;
+import com.loopers.interfaces.api.member.dto.MemberV1Dto;
 
 /**
  * Member 관련 API 입니다.
  */
 public interface MemberV1ApiSpec {
-    ApiResponse<MemberRegisterResponse> register(MemberRegisterRequest registerRequest);
-    ApiResponse<MemberInfoResponse> find(Long memberId);
+    ApiResponse<MemberV1Dto.MemberRegisterResponse> register(MemberRegisterRequest registerRequest);
+    ApiResponse<MemberV1Dto.MemberInfoResponse> find(Long memberId);
 }
