+++ apps/commerce-api/src/test/java/com/loopers/domain/MemberFixture.java
@@ -1,6 +1,9 @@
 package com.loopers.domain;
 
-import static com.loopers.domain.Gender.MALE;
+import static com.loopers.domain.member.Gender.MALE;
+
+import com.loopers.domain.member.Member;
+import com.loopers.domain.member.MemberRegisterRequest;
 
 public class MemberFixture {
     public static Member createMember() {
