+++ apps/commerce-api/src/test/java/com/loopers/domain/MemberIdTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
+import com.loopers.domain.member.MemberId;
+
 class MemberIdTest {
     @Test
     void equalityTest() {
