+++ apps/commerce-api/src/test/java/com/loopers/domain/MemberTest.java
@@ -6,6 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import org.junit.jupiter.api.Test;
 
+import com.loopers.domain.member.Gender;
+import com.loopers.domain.member.Member;
+import com.loopers.domain.member.MemberRegisterRequest;
+
 class MemberTest {
     @Test
     void createMemberTest() {
